Dockers-etc/ELK/shared/python-elk/requirements.txt
Flask==2.0.1
python-logstash-async==2.3.0